login and sign up
I kind of changed everything you had maggie sorry. For users I needed to implement sth called forms and to do so I had to change the htmls and css. Now they are working but we need to implement the store views. i added in the path so that after login in it will redirect the user to there.